Ignore symlinks in the tree walk
Trying to open a symlink directly doesn't go well, and may reference
things outside the repository.  Assuming that a symlink doesn't need to
be independently opened and checked as if the file is in the repository
it'll get picked up elsewhere, and this would just be a dupe, or if it's
external it's not part of the repository and not entirely useful to
check anyway.

If it's a symlink, just skip it.

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>